Fix exitHandler to handle empty containers array

+++ src/FhenixHardhatRuntimeEnvironment.ts
@@ -74,6 +74,10 @@ function randomBetween(min: number, max: number): number {
 // exitHandler blocks, so that the process won't exit before the containers are removed
 // note: `docker rm -f` gracefully skips non-existing containers
 function exitHandler() {
+  if (containers.length === 0) {
+    return;
+  }
+
   child_process.execSync(
     `docker rm -f ${containers.map((name) => `"${name}"`).join(" ")}`,
   );
